pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "readmypaper"
version = "0.2.0"
description = "Local scientific PDF-to-audio reader with AI-driven cleanup and a web UI"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "GPL-3.0-or-later"}
authors = [{name = "ReadMyPaper Contributors"}]
keywords = ["pdf", "audio", "tts", "text-to-speech", "scientific", "papers", "accessibility"]
classifiers = [
  "Development Status :: 4 - Beta",
  "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Operating System :: MacOS",
  "Operating System :: Microsoft :: Windows",
  "Operating System :: POSIX :: Linux",
  "Topic :: Multimedia :: Sound/Audio :: Speech",
  "Topic :: Scientific/Engineering",
]
dependencies = [
  "fastapi>=0.115,<1.0",
  "uvicorn[standard]>=0.34,<1.0",
  "jinja2>=3.1,<4.0",
  "python-multipart>=0.0.9,<1.0",
  "platformdirs>=4.2,<5.0",
  "langdetect>=1.0.9,<2.0",
  "docling>=2.84,<3.0",
  "piper-tts>=1.4.2,<2.0",
  "httpx>=0.28,<1.0",
  "numpy>=1.26,<3.0",
]

[project.urls]
Homepage = "https://github.com/ThalesMMS/ReadMyPaper-dev"
Repository = "https://github.com/ThalesMMS/ReadMyPaper-dev"
Issues = "https://github.com/ThalesMMS/ReadMyPaper-dev/issues"

[project.optional-dependencies]
kokoro = [
  "kokoro>=0.7,<2.0",
  "soundfile>=0.13,<1.0",
]
dev = [
  "pytest>=8.2,<9.0",
  "ruff>=0.9,<1.0",
]

[project.scripts]
readmypaper = "readmypaper.main:cli"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["readmypaper*"]

[tool.setuptools.package-data]
readmypaper = ["templates/*.html", "static/*"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["B008"]
